+++ src/models/demand.ts
@@ -1,0 +1,6 @@
+import { ACTION } from './priv-terms.js';
+
+export interface Demand {
+  action: ACTION;
+  message?: string;
+}
