+++ src/models/privacy-request.ts
@@ -1,0 +1,5 @@
+import { Demand } from './demand.js';
+
+export interface PrivacyRequest {
+  demands: Demand[];
+}
